requirements.txt
xlrd==1.2.0
Pillow==8.1.2
PyQt5==5.15.4
